fix: multi-line tag message when using -m
Gitflow release finish and hotfix finish allow for a tag message to be
entered from the command line. Currently only the last messeage is read
by git flow if multiple messages are passed in on the command line.

The git binary does not allow for the use of the \n escape sequence when
entering a message from the command line via -m. As a result multiple
invocations of the option must be used for a multi-line commit message.

As a workaround when using the git binary directly we can use printf
inside of the commit message to properly handle \n escape codes. This
allows for a single invocation of -m. To use this workaround inside of
the double quotes we use $(printf 'line of text with \n escape codes
placed as needed for proper formating of the message')

This commit will add the above workaround when entering a tag message
from the command line for release finish and hotfix finish.