feat(feature): add release sub command
Add sub command to git flow feature that allows for a release to be
created based off the current feature branch. When running feature
release the feature branch is rebased with develop prior to the creation
of the release branch. When finishing the release the resulting tag is
back merged to the feature branch and not onto the production or develop
branches.

resolves #63